[PATCH 2/9] qbsp: Replace WAD class
Replace the WAD class with plain C code. Class WAD becomes wad_t and the
member functions are just normal functions taking a wad_t as an argument.

The iWad member was a bit strange. It was used in some member functions for
walking the wadlist array. I've elimanted that from the struct and made the
functions track that state themselves.

Signed-off-by: Tyrann <tyrann@disenchant.net>