lightpreview: fix texture mips not being generated
This was happening because allocateStorage allocates space for the texture and its mips up-front, the user of QOpenGLTexture is required
to determine the number of mips required for a texture beforehand.
allocateStorage was allocating space for mip 0 and no others, leading to no additional mips being generated.

This also adds explicit format and type to allocateStorage for only this instance, although it still remains optional.